Deleted slots and history page, cleaned routes
Deleted the slots pages because we don't have a data model for it. Also
deleted the history page because it's pretty much a duplicate of the
auction page. I also moved the content of history.ejs over to the
auction.ejs file.

The routes were starting to become a mess so I've cleaned them up and
separated them to be clearer to read. All api routes without the prefix
where also fixed and all the links were updated as well. I've created a
doc that outlines the exact specifications of each route as well.