Auction Detail, history pages, db fixes, misc fixes
Filled out the Auction Details and History Pages along with all
associated backend references as well. The Auction details page also has
all the bids for that auction added to the page as well.

Made all the DB calls return the value of the key-value pairing.

Changed appendBidsToAuction db call so that it returns the origin
auction object instead of the one returned with the call.

Changed the Admin page so that the details link in the open and closed
sections are in the same spot as the rest of the sections.

Added the Auction id to the index page.

Removed the last links to the deprecated slots page in all the other
pages.

Finally, I added back in the user parameter in the history template that
I removed by accident in the last commit.