Added update auction and bids, misc fixes
Added the SB and api routes to update both auctions and bids. This
includes all the backend work to makes these calls work. I also added a
lot of validation to the calls to ensure bids can't be sneaked in after
an auction is closed, change another user's bid, etc. There is more
validation needed.

I also added more validation to creating auctions but more could be done
here as well.

Fixed an issue where the DB call getBid would hang the server if an id
that's not a bid ID is given.

Fixed the way usernames are pushed into the bid on insertion. I
basically renamed the var being passed in from username to user because
an object is passed in and not just the username.

I also went through all the SB routes and api routes and separated out
all the inconsistencies with the way we were using some of the routes. I
also created mirror SB routes for all the api routes. This allows each
route to exit differently. SB routes redirect to the appropriate page
while the api routes always return some kind of json.