update auction/bid validation, auction labels, etc
Completed the validations for both updating and creating auctions and
bids in the DB layer. For creating auctions and bid, this includes
making sure that all inputs exist are given, making sure that those
inputs are valid numbers or boolean values, that start time is always
before end time, and that slots is not negative. For updating auctions
and bid, any inputs that are passed in are checked for type validity,
start time is before end time, and that slots are not a negative
number.·

I also made sure that for the updating calls, only the values given are
updated on the db. Any values not given retain their original values.

I updated the auction details page so that the appropriate label is
given at the top of the the page (Open, Closed, Upcoming, and Past).

Lastly, I updated the bidding controls in the sbindex page to use the
highest bid price + 0.05 BTC as the default value.