Auction Edit Page, plural api call, auction history
Added the Auction Edit Page that allows a user to edit a given auction.
It uses the same validation as the auction creation form in the admin
page. This includes all backend work to grab an auction and display it
through the new routes. This routes should be secured so that only
admins can use it.

Renamed all view and api urls to have a plural resource collection
naming scheme. (Ex. auction -> auctions).

The Auction History page now shows auctions from the past rather than
"closed" auctions that are still running but disabled by an admin.