Delete Auction and Bid (views/api)
Added the API and SB routes for deleting auctions and bids. This
includes all the backend work to get this to work including the DB calls
and Express routes.

There's currently no view for deleting a bid and these routes should be
locked down so only admins can use these calls.

A delete auction button was added to the admin page to allow the user to
delete specific auctions. Additional buttons were added to all auction
categories to allow the user to edit or delete auctions easily.