JS rounding error, bid history, auction naming
Fixed the rounding error problem in the bidding controls on the index
page. Occasionally it would run into a rounding error giving extremely
long decimal values. Now it's rounds to the hundredth place.

The Auction details page now includes the bid history under the list of
winning bids. On the index page, the auction ID is linked to the details
page of that particular auction.

Renamed the Auction categories on the admin page to better reflect the
type of auction being shown. (Closed Scheduled Auctions -> Disabled Open
Auctions, Open Scheduled Auctions -> Open Auctions)