DB calls for ad, Ad model, ad view/api routes
Created basic DB CRUD operations for ads in db.js. This includes some
basic validation but not the full HTML validation. Also created the ad
model for emcee to get a specific ad. Lastly, created all the CRUD view
and api routes for the router. The views have not been created yet.