Ad views, and User's ads
Re-wrote the profile view to show the full breath of ads for a
particular user. This includes approved, submitted and saved ads for a
single user. Controls to edit or delete an ad where also added to each
of the sections.

This page also includes a link to a another page where the user can
create or edit ads. There is also another link that shows the public
view of a single ad. This page when viewed by an admin, has controls to
approve or reject an ad for use.

This work cannabalizes the origin ad upload page and uses it to create
and/or edit ads now. This work also includes all backend work to support
getting all ads for a user, extra post methods for ads,  as well as the
initial work for admin functions for ads.

Extra SB routes were created to support ad approval, ad rejection, and
ad deletion via the POST HTTP method. This is to support noscript is all
user facing pages and operations.

Lastly, the layout was edited and the "My Ads" link was removed. The
Profile and the Admin links were edited so that they only appear if a
user is logged in. Otherwise only the Home and History links appear in
the sidebar.