Admin enforcement, card layouts, html validation fix
Admin view and api routes are now protected. Any route admin related
checks the user object in the request for an admin property. If this
property isn't found, the user is automatically redirected or any
protected properties are not updated in the DB.

Created a card layout for ads that is used in both the ad page and the
profile page for users. Other UI elements were touched up. The elements
updated were the log out button, sidebar icons, history page, profile,
auction and admin pages.

Finally there was an issue with the name of the validation module being
used in the DB wrapper. The wrong name was being used. It was updated to
the correct name and usage.