Anti-Bid Sniping, registration code clean up
Modified the create and update auction DB functions to generate a random
end time that is saved as trueEnd. Cleaned all DB function and view/API
routes to ensure that the trueEnd property is not exposed to the public
in any way.

Added the Anti-Bid sniping verbage to the auction details
page.

The algorithm for the Anti-Bid sniping feature basically chooses a
random millisecond at some point 30 mintues after the stated end time
for an auction. This millisecond is chosen by the equation:

Math.floor(Math.random() * ((1000 * 60 * 30) + 1)) + end time (ms)

The code was also modified so that all variants of the get auction db
function also calculate and return the auction.open property. Any future
code involving auctions should rely on this property to check if the
auction is open or not.

Cleaned up and removed any superfluous code relating to auctions in the
registration view/backend files.