Submitted Ads Page
Created the submitted ads page that lists all the ads submitted by all
users on one simple page. This includes routing for both the view and
api routes, the model to handle this call, the DB call and a new couch
view that specifically searches for these ads. Currently the couch view
simply filters all ads with the submitted property set to true.

I also added a nav bar and link to the submitted ads page in the admin
page for easy access.