Auction closing: the end, heckler, email templates, dates
Finished all the work required to complete the auction closing
notification feedback loop. This includes the email templates for both
winners and bidders. The winner's email templates include the BasicPay
invoice generated by calling the POST /invoices on BasicPay. The
Bidder's template gives the date and time for the next available
auction in UTC.

Heckler was integrated into the auction closing feedback loop to send
out the emails to all the winners and the bidders. The configurations
for heckler are set in the .env file.

Implemented date formatting for both Javascript users and non-Javascript
users. Each template that loads a date, outputs the date as a UTC
String. So users with noscript will see this date. Users that have JS
turned on will also activate a script that turns that UTC date string
into a local time string.