Registraion module and email template, invoice module
Started work on the registration feedback loop. Edited the auctions to
have a list of registered users and edited the db create and update
calls to support this. The auction view was also updated to support
registering per auction and displaying the registration status for users
already registered.

The registration task has been split out into its own module as well as
the invoicing module. The winner's task is not using this invoice module
yet. The registration email template was also written to be used with
the registration task.

Each auction displays a button to register the current user with this
auction. Upon clicking the button, the user is added to the current
auction object with a registration status and registered property. An
invoice for the registration fee is created along with a BasicPay
Receipt and the link to the invoice is emailed to the user.

Updated the site's routes and seed file to support registrations.

Webhooks for both the winner's invoice and the registration invoice are
still needed as well as fixing the redirect routes for registration. The
registration view needs to be updated as well.