One Time Registration and registered users
Updated the Registration Feedback loop so that it's a one time
registration fee instead of a fee per auction. This included updating
adding a registeredUser object into the DB each time a user starts the
registration process. The DB calls to upsert and retrieve
registeredUsers were also create. A Emcee model was created to easily
and consistently retrieve the registeredUser object from the DB

The registration email template, auction view, and registration views
(error and success), were updated to reflect the changes above.

Modifications to the registration and invoice modules were also
necessary due to the way the invoices for registration fee are now
different. A auction isn't reference in the invoice so that has been
pulled out of those modules.

The auction, bid, and registration routes have all been updated to check
for a registeredUser for some of their actions. Auctions will not show
bidding controls if the user is not registered but will show either a
registration status or the option to register if they have not already.
The bid route will now reject bids if the user is not registered. The
registration route will only allow a user to register once and redirect
to a instructions page or otherwise to an error page showing the user a
relevant error ("User already Registered", "No User Found", etc.)

Lastly, the registration route was moved from /sb/auction/registration
to just /registration.