Registration and notify winners webhook
Created the webhooks for Basic Pay to call when the payments for both
registration and auction winner are verified. In both cases, an email is
sent to an admin stating that the payment for either case has been made.
In the registration case, the user is also allowed to make bids on
auctions. Email templates for both cases were also created.·

The BPReceipt object in the DB needed to be modified to use a user's
userId instead of their username. The same change needs to be done for
both ads and bids.·

A few more configurations were added, mainly the base URL for the site
itself to be used in both emails to the admin as well as for the
webhooks in BasicPay.