userId, admin email, auction closing
Renamed the cull-auctions resque job to auction closing to better
describe what it's actually doing. Also, refactored the code in auction
closing to use the invoice module instead of having its own version of
that code.

Removed all instances of hard coded emails from the code. Made the admin
email address and the mailer sender email address a configuration
option. Updated the code to use these configuration options instead of
the hardcoded addresses.

Updated both the ad and bid DB models to use the userId as the key
constraint rather than using the username. The reason for this is
because usernames are trasient in some system and can be changed. UserId
should be a value that is permanently assigned to a user.

Lastly, removed the admin property on the user object before it is saved
in the bid DB object model. Although this isn't a security risk, it's
just the right thing to do.