GeoLocate IP and view ads endpoint
closes #9

Added the GeoIP-lite node module to accept and geo locate IPs sent to
the view ad endpoint. This will filter out all ads that are not allowed
in the origin country of the IP. The view ad endpoint also takes in a
limit parameters that returns an array of location specific ads where
the length of the array matches the limit parameter.

All ads returned by the endpoint are chosen at random after the location
rejected ads are pulled out. Output from this endpoint is always an
array of ads in json format.