Check for DB on startup
closes #14

Added a check on startup to see if the couchDB DB exists. If it doesn't
exists, the startup script builds the DB and pushes the couchdb ddoc and
views to the newly created DB. If the DB cannot be started, then the
startup script will gracefully exit with an error.

This not only closing the issue with the app crashing if the DB doesn't
exist but goes beyond that to create the DB if it can.