Removed hardcoded DB name
Removed all instances of the hardcoded 'adness' db name from throughout
the code base. Also updated the seed file to use the same code as the
startup script to create a db if it doesn't exists already. This
includes loading the ddoc and views into the newly created db as well.

The db name is only referenced in the config file and is propagated
throughout the code base now.