Update BasicPay to Baron, Login Rate Limiter
Updated both the package.json and the code base to use the new Baron
name instead of BasicPay. This includes updating the config variable and
its references.

Included a rate limiter on the login route that is set to 10 tries over
a one minute period. This limiter logs IPs that exceed the rate limit to
be banned at a later point in time.