Fixed log output, redesign, admin registration, ip banning
Edited the log output to not be so verbose. Originally, the application
was logging all resque output including worker polling and pausing. This
is too verbose and would clog up the logs.

Added a font and background pattern to help with the aesthetics of the
views.

Added an extra line in the code to allow admins to be immediately
registered immediately without paying. This is for testing purposes only.

Failed logins are redirected to route "/failedLogin." This was done to
capture when failed login attempts are made and with which IPs. After
enough failed attempts are made, the IP will be banned from the
server completely. IP Banning is done outside of this application using
another software.