Admin config, lowest bid on auction
Changed the way admins are defined in the configuration. Instead of
using usernames, userId are used instead to identify if a user is an
admin or not. Usernames are not permanent and could change but userId
are forever.

Changed the way the latest price is calculated on the auction bidding
page (/auction/:autionId). Instead of finding the highest bid price
possible, it now finds the lowest price to beat.