Auction sorting, create auction time bug
Auctions on the history, index, and admin pages are now sorted by either
auction start time for most cases and auction end time for past or
history auctions.

There was a bug with creating new auctions where 12 AM was interpreted
as being 12 PM. Adjusted the JS code in sblayout to interprete it
correctly.

closes #17