Adjusted Invoice form to display payments per slot
Changed the way the invoice from being sent to Baron is created so that
each line item sent to Baron corresponds to a ad slot that was bid on.
Originally, each line item was the remainder of the total payment
divided by the number of slots. This doesn't account for bids of varying
prices.

closes #18