Various UI fixes
Added a border around the textarea in the ad_upload page. Changed the
way the buttons were shown on that page. Used a button groups to fill
the width of the section.

Added space between buttons on the ads view page. Also fixed all the
redirects so the user is redirected to their profile page instead of the
index page where appropriate.