Exhaustive ad randomization, admin controls in ads
Fixed a bug where the reject button of the admin controls for the ad
view page is visible to the public. The button did not work even though
it was visible. It was a screwup in the template logic.

Modified the Ad Randomization algorithm to use an exhaustive search of
the winner's ads rather than an random selection. This guarantees that
when the number of ads requested equals the total number of slots
sold through an auction, each winner's number of slots is accurately
represented. For calls to the algorithm with a requested number of ads
less than the total sold, an exhaustive search occurs but all winners
may not be equally represented. For calls to the algorithm where the
requested number of ads is larger than the number of slots sold, the
winners are guaranteed to have their number of slots bought represented,
and any extra are culled through another iteration of the exhaustive
search but is not guaranteed an equal distribution.