Webhook bug, min confs config
Fixed an issue with the auction payment webhook sending out the wrong
url to Baron. The url sent out has been fixed to point to the right url.

Added another config option to set the number of minimum confirmations
to signify that a payment is valid. This number is sent to Baron.