Data leak in registration webhook
There's a potential data leak in the registration webhook when a valid
call is made twice within a short period of time. The error that is sent
back in the second call has information that shouldn't be made public.

The error message has been modified to a simplier status code 500 with
the error's message instead of the full json error object.