Fixed Issue #22 Ads needed to bid
Fixed the issue where users are allowed to bid even though they have no
ads in rotation. Adjusted the gate that blocks users from being able to
bid to also check to see if the current user has at least one ad that
has been approved and put into rotation.

Also adjusted the Registration button's verbiage to clearly state that
the registration is for the system as a whole and not on a per auction
basis.

closes #22