Removed SHA256, payment status validation on webhooks
Removed the SHA256 handshake on webhook calls between Adness and Baron.
This includes reversing all the backend work to make the SHA256
handshake work such as the couchapp design doc addition to recall BP
Receipts by token and the associated DB call. Lastly, the actual hashing
during the webhook and invoice creation calls were removed.

In place of the SHA256 handshake, upon calling the webhook, Adness goes
and queries Baron itself for the status of the invoice and payment. If
the status returned is set to paid, then Adness continues on with
emailing the admin saying the payment has been made.