Added a notification to user on payment received
Emails are now sent to the user when a payment for either registration
or auction winning are paid out. For registration, the user is also
given basic instructions on how to create an "in rotation" ad so they
can start bidding. For Auction's won, those emails also have basic
instructions to double check that they have "in rotation" ads ready to
be used.

On Warren's suggestion, we should come back and look at this to add
price and USD value at time of invoicing to the auction's won emails at
a later date.