Invalid Invoice retry queue, admin auction dates bug
Fixed a bug with the auction creation tool on the admin page that
doesn't recognize when a user chooses a date in the months ahead or
behind the current month.

Created a invoice retry queue in the scenario where Baron is down. The
new code will save the invoice and receipt as an object to the DB so
that a continually running interval process can pick up that object and
tries to create an invoice with baron at a later time.

The queue itself is smart enough not to run again until the current
interval is fully completed. It also tries each queued invoice one at a
time to keep resources down. A side effect of this is, the time between
intervals are not exactly a fixed time length but will rather wait until
the current interval is complete, then wait the designated time length
before starting the next interval.

An invoice is queued when Baron cannot be reached or any other
connection type error is found during the request call to Baron. After a
queued invoice is successful, it is deleted from the queue and therefore
the DB as well.

closes #23