Bid creation time bug, links in ad validation bug
Fixed a bug with the format of the creation time of a bid on the auction
view. It was still using the old format. The new format have been
applied to the auction view.

Fixed a bug with the server side ad validation. It was removing links
from the ad but has been updated now to allow links using the https
protocol.