Updated ad, user, submittedAds pages, bug fixes
Bug Fixes:
Fixed a bug where db.getRegisteredUser could potentially return the
wrong user. This is due to how couchdb uses start and end keys as the
search criteria. Originally it was left open ended with {} but has since
been changed to user the same value as the start key so that it only
returns one object.

Fixed a bug where the users page would use the information of the
current viewing user instead of the user that the page was created for.
This is most apparent when viewing the ads for user foo and having the
username for user bar (who is currently viewing the page) show up.

Fixed a bug where rejected ads would still appear to be rejected even
after being edited and re-submitted. After a rejected ad was edited and
submitted, the status logic was not reset so although the ad was changed
and the boolean for submitted was set to true, the boolean for rejected
was left to true as well. Rejected is higher in priority to submitted
and thus the edited ad would still appear to be rejected. The boolean
are now reset upon updating.

Changes:
Updated the profile (users) page to show a link to a user's posts on the
forum.

Updated the ad page to show a link to both the ad owner's page and the
posts for the ad owner on the forums.

Updated the ad partial view to show region information for that
particular ad.

Updated the submittedAds page to show a link to the user per ad and the
ad's user's posts on the forum. Also wrapped each ad in it's own panel
rather than one panel for all ads. This is visual cleaner and allows for
approve and reject buttons for each ad in the panel footer. The
submittedAds page can now handle approving/rejecting mass ads from just
one page. This required using a bit of JS so now the submittedAds page
requires JS as well.