Bug Fixes, more user links
Bug Fixes:
Fixed a bug where the current user's ads weren't getting retrieve on the
auction page. This is due to a change made in the last commit and how
the userAds model was getting the user's id. The auction page's js now
adjusts for this change.

Changes:
Changed the wording on the auction page to be more clearer.

Changed all the usernames in the bid section of the auction page to be
links that link to the bidder's ad page.