Auction description
Added a text field that allows HTML but strips it of any malicious
content to the auction. The db.js file, the auction.js route file, and
the front files needed to be edited to add this new property.

It is currently shown in the auction partial view and thus can be found
in all iterations of the auction.

closes #36