Auction Regions
Initial implementation of regions for auctions:
Added a region field to the auction. This includes updating the
databases functions (create, update) to handle the new field, the
auction route functions (update), and the front-end JS to pass the
region information to the backend.

A new validation function was created to validate the regions for
auctions. As well as, a new config field that allows the admin to create
whitelisted regions to be used. A region drop down was added to the
auction creation and update tool in both the admin and auctionEdit
pages. And the auction partial view was updated to show the region
information. Also, the history page and the admin page was updated to
region information at a glance.