Initial work on new region system for auctions
Implemented the new region object config in the config.js file. There is
now only one config variable for regions and it is an array of region
objects where each region has a name, an array of countries, and a
inclusive boolean.

Removed the top level slots property from the auction model. This was
removed in the validation.js file, the db.js file, the auction.js file,
the admin page, and the sblayout.ejs file. It still needs to be fully
removed from the auction edit page.

Created new validations for regions in both the front end js and in the
validation.js file. This validates that there is at least one region,
the region names are found in the config, and that the slots property in
each region is a valid number and is greater than zero.

Updated the <hr> dom element to be darker so it shows up on the page
better.

Optimized the time format code in the auction.js file and pulled in all
the region names from the config to be display on the admin page.

Also pulled the region names from the config into the auction.js file's
editAuction method so that that AuctionEdit page can also show all the
region names.

Added a region control to the auction creation tool on the admin page.
This needs to be replicated on the AuctionEdit page.

Updated the submit and update methods on the admin.ejs and sblayout.ejs
file respectively to reflect the above changes. (Removed slots and added
regions/region validations.)