Added ads start and end times, AuctionEdit updates
Syntax update in config file.

Added start and end times to the auction model. Created new validations
for the ads start and end time that also validate that the ads start
time is after auction end time. Also fixed a bug with the region
validation trying to validate that the region.slots is a number.

Added ads start and end times to the auction.js route file.

Updated the admin.ejs page with ads start and end datetimes. Refactored
the auctionEdit.ejs file to match the new auction creation tool on the
admin page.

Updated all the front end JS to account for and validate the new ads
start and end datetimes.

Loaded the new auction region and auction ads start and end datetimes
into the newly updated auction creation tool on the auctionEdit.ejs
file.