regions in bids, view updates, auction probabilities
Updated the bidPerSlot in the bidding algorithm to primarySlots. This
was changed in both the bidding algorithm and the db methods that use
the algorithm.

Updated the bid model to include a region. This affected the db.js file,
and the bid route files. A validation was also created to validate that
the region given is in the auction.

Updated the auction and sbindex views to with a bid per region tool.
Also optimized the auction view to need fewer variables to be passed
into the template. This includes splitting the auction partial view into
two different files for both the auction and the sbindex page. The one
used in the auction page includes the bid tool while the other doesn't.

Updated the auction.js and template file to find the lowest bid price
for all regions and attached that to the region object.

Created a file to figure out the probabilities for all slots in every
region of an auction. This file has been included and used in both the
sbindex and the auction page.