Delete regions from auction
Added the ability to remove a region from an auction as long as that
region does not already contain bids. This involves updating the UI on
both the admin and the auctionEdit pages.It required adding the bid
model to the auction.js route file and checking the regions in each bid.
If an auction region has a bid and the that region is deleted in the
input, that region is added back to the input before the auction is
processed.

This makes sure the validation exists in both the ui and the route but
not in the db file. The API routes will need to be updated to account
for this.