Regions in ads
Updated the ad model to use the new region design rather than just the
blacklisted regions. This involved updating the db.js file to define
regions in the ad model, the ads.js route file to remove the old
blacklistedCN and add the region property, and the front end views
(ad_upload.ejs and ads.ejs).