Regions in auction close and bug fixes
Added region to the auction close event. This includes adding regions to
the AdsInRotation object (otherwise known as the reference object).
Re-routing the way winners are found over all the regions, while also
keeping the winners for each region, in the region object. Also added
regions to the lineitems for invoice as well so each lineitem in the
invoice can display which region it's for.

Fixes:
Fixed a bug where an in rotation ad still appears to be in rotation
after it is rejected. The inRotation Boolean is now reset on rejection.

Fixed the was Ads Start and End times are kept in the DB and eventually
displayed to the user. It was being stored as a string, but is now a
number in epoch time. This lead to a bug where Ads Start and End time
appeared invalid for auctions in the past.

Added an extra object check on the ad_upload page. This page broke on
new ads because it was looking for an ad that did not exist. It makes
sure that if there is no given ad, it gives up and the rest of the
checks.

Fixed a bug on the history page where the regions were not showing up
correctly. The regions are now shown as a list of region instead of as
an array of objects.