Added a alreadyPaid check to registration webhook
Added a variable called alreadyPaid that checks to see if a user has
already paid his registration fee. This check bypasses the rest of the
webhook process so if the system gets multiple webhook calls from baron,
it will not keep sending emails to the user saying that they've paid the
registration fee.