Profile Page bug for new users
New user were greeted with a page that showed "No User Found" and no
controls to create ads on the profile page. Now users are greeted with
a full and intact header that show the username, posts, and a link to
create ads. But are also show a small panel that asks the user to
register with the system as well.