Fixed a bug with the bidding algorithm and second slots
Fixed a bug where a bid has only some of its slots in the winning area
does not have the rest of its slot represented in the potential winners
area. This was fixed by not deleting the bid if there are still slot
leftover and adding to the potential winners area right before running
the loop for the potential winners.